+++ cli/src/common.rs
@@ -1,8 +1,6 @@
 use anyhow::Result;
-use base64::engine::general_purpose;
-use base64::Engine as _;
 use bonsol_prover::input_resolver::{ProgramInput, ResolvedInput};
-use bonsol_sdk::instructions::{CallbackConfig, ExecutionConfig};
+use bonsol_sdk::instructions::CallbackConfig;
 use bonsol_sdk::{InputT, InputType, ProgramInputType};
 use clap::Args;
 use rand::distributions::Alphanumeric;
@@ -50,7 +48,7 @@ pub struct ZkProgramManifest {
 #[serde(rename_all = "camelCase")]
 pub struct CliInput {
     pub input_type: String,
-    pub data: String, // base64 encoded if binary
+    pub data: String, // hex encoded if binary with hex: prefix
 }
 
 #[derive(Debug, Clone)]
@@ -91,12 +89,20 @@ impl FromStr for CliInputType {
 #[serde(rename_all = "camelCase")]
 pub struct ExecutionRequestFile {
     pub image_id: Option<String>,
-    pub execution_config: ExecutionConfig,
+    pub execution_config: CliExecutionConfig,
     pub execution_id: Option<String>,
     pub tip: Option<u64>,
     pub expiry: Option<u64>,
     pub inputs: Option<Vec<CliInput>>,
     pub callback_config: Option<CliCallbackConfig>,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct CliExecutionConfig {
+    pub verify_input_hash: Option<bool>,
+    pub input_hash: Option<String>,
+    pub forward_output: Option<bool>,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize, Default)]
@@ -193,15 +199,19 @@ pub fn execute_transform_cli_inputs(inputs: Vec<CliInput>) -> Result<Vec<InputT>
         let input_type = CliInputType::from_str(&input.input_type)?.0;
         match input_type {
             InputType::PublicData => {
-                if is_valid_base64(&input.data) {
-                    let data = general_purpose::STANDARD.decode(&input.data)?;
-                    res.push(InputT::public(data));
+                let has_hex_prefix = input.data.starts_with("0x");
+                if has_hex_prefix {
+                    let (is_valid, data) = is_valid_hex(&input.data[2..]);
+                    if is_valid {
+                        res.push(InputT::public(data));
+                    }
+                    continue;
                 }
                 if let Some(n) = is_valid_number(&input.data) {
                     let data = n.into_bytes();
                     res.push(InputT::public(data));
+                    continue;
                 }
-
                 res.push(InputT::public(input.data.into_bytes()));
             }
             _ => res.push(InputT::new(input_type, Some(input.data.into_bytes()))),
@@ -210,21 +220,19 @@ pub fn execute_transform_cli_inputs(inputs: Vec<CliInput>) -> Result<Vec<InputT>
     Ok(res)
 }
 
-fn is_valid_base64(s: &str) -> bool {
+fn is_valid_hex(s: &str) -> (bool, Vec<u8>) {
     if s.len() % 4 != 0 {
-        return false;
+        return (false, vec![]);
     }
-    let is_base64_char = |c: char| c.is_ascii_alphanumeric() || c == '+' || c == '/' || c == '=';
-    if !s.chars().all(is_base64_char) {
-        return false;
+    let is_hex_char = |c: char| c.is_ascii_hexdigit();
+    if !s.chars().all(is_hex_char) {
+        return (false, vec![]);
     }
-    let padding_count = s.chars().rev().take_while(|&c| c == '=').count();
-    if padding_count > 2 {
-        return false;
-    }
-    general_purpose::STANDARD.decode(s).is_ok()
+    let out = hex::decode(s);
+    (out.is_ok(), out.unwrap_or_default())
 }
 
+#[derive(Debug, PartialEq)]
 pub enum NumberType {
     Float(f64),
     Unsigned(u64),
@@ -243,19 +251,26 @@ impl NumberType {
 }
 
 fn is_valid_number(s: &str) -> Option<NumberType> {
-    if let Ok(num) = s.parse::<f64>() {
-        return Some(NumberType::Float(num));
-    }
     if let Ok(num) = s.parse::<u64>() {
         return Some(NumberType::Unsigned(num));
     }
     if let Ok(num) = s.parse::<i64>() {
         return Some(NumberType::Integer(num));
     }
+    if let Ok(num) = s.parse::<f64>() {
+        return Some(NumberType::Float(num));
+    }
     None
 }
 
-fn parse_entry(index: u8, s: &str) -> Result<ProgramInput> {
+fn proof_parse_entry(index: u8, s: &str) -> Result<ProgramInput> {
+    if let Ok(num) = s.parse::<i64>() {
+        return Ok(ProgramInput::Resolved(ResolvedInput {
+            index,
+            data: num.to_le_bytes().to_vec(),
+            input_type: ProgramInputType::Private,
+        }));
+    }
     if let Ok(num) = s.parse::<f64>() {
         return Ok(ProgramInput::Resolved(ResolvedInput {
             index,
@@ -270,24 +285,19 @@ fn parse_entry(index: u8, s: &str) -> Result<ProgramInput> {
             input_type: ProgramInputType::Private,
         }));
     }
-    if let Ok(num) = s.parse::<i64>() {
-        return Ok(ProgramInput::Resolved(ResolvedInput {
-            index,
-            data: num.to_le_bytes().to_vec(),
-            input_type: ProgramInputType::Private,
-        }));
+    let has_hex_prefix = s.starts_with("0x");
+    if has_hex_prefix {
+        let (is_valid, data) = is_valid_hex(&s[2..]);
+        if is_valid {
+            return Ok(ProgramInput::Resolved(ResolvedInput {
+                index,
+                data,
+                input_type: ProgramInputType::Private,
+            }));
+        } else {
+            return Err(anyhow::anyhow!("Invalid hex data"));
+        }
     }
-    if is_valid_base64(s) {
-        let decoded = general_purpose::STANDARD
-            .decode(s)
-            .map_err(|e| anyhow::anyhow!("Error decoding base64 input: {:?}", e))?;
-        return Ok(ProgramInput::Resolved(ResolvedInput {
-            index,
-            data: decoded,
-            input_type: ProgramInputType::Private,
-        }));
-    }
-
     return Ok(ProgramInput::Resolved(ResolvedInput {
         index,
         data: s.as_bytes().to_vec(),
@@ -302,7 +312,7 @@ fn proof_parse_input_file(input_file: &str) -> Result<Vec<ProgramInput>> {
             .inputs
             .into_iter()
             .enumerate()
-            .flat_map(|(index, input)| parse_entry(index as u8, &input.data).ok())
+            .flat_map(|(index, input)| proof_parse_entry(index as u8, &input.data).ok())
             .collect();
         if parsed.len() != len {
             return Err(anyhow::anyhow!("Invalid input file"));
@@ -325,8 +335,7 @@ fn proof_parse_stdin(input: &str) -> Result<Vec<ProgramInput>> {
             '}' | ']' if !in_quotes => in_brackets -= 1,
             ' ' if !in_quotes && in_brackets == 0 && !current_entry.is_empty() => {
                 let index = entries.len() as u8;
-                println!("{}", current_entry);
-                entries.push(parse_entry(index, &current_entry)?);
+                entries.push(proof_parse_entry(index, &current_entry)?);
                 current_entry.clear();
                 continue;
             }
@@ -335,7 +344,7 @@ fn proof_parse_stdin(input: &str) -> Result<Vec<ProgramInput>> {
         current_entry.push(c);
     }
     if !current_entry.is_empty() {
-        entries.push(parse_entry(entries.len() as u8, &current_entry)?);
+        entries.push(proof_parse_entry(entries.len() as u8, &current_entry)?);
     }
     Ok(entries)
 }
@@ -347,4 +356,102 @@ pub fn rand_id(chars: usize) -> String {
         .take(chars)
         .map(char::from)
         .collect()
+}
+
+#[cfg(test)]
+mod test {
+    use super::*;
+
+    #[test]
+    fn test_proof_parse_stdin() {
+        let inputs = r#"1234567890abcdef 0x313233343536373839313061626364656667 2.1 2000 -2000 {"attestation":"test"}"#;
+        let inputs_parsed = proof_parse_stdin(&inputs).unwrap();
+
+        let expected_inputs = vec![
+            ProgramInput::Resolved(ResolvedInput {
+                index: 0,
+                data: "1234567890abcdef".as_bytes().to_vec(),
+                input_type: ProgramInputType::Private,
+            }),
+            ProgramInput::Resolved(ResolvedInput {
+                index: 1,
+                data: "12345678910abcdefg".as_bytes().to_vec(),
+                input_type: ProgramInputType::Private,
+            }),
+            ProgramInput::Resolved(ResolvedInput {
+                index: 2,
+                data: 2.1f64.to_le_bytes().to_vec(),
+                input_type: ProgramInputType::Private,
+            }),
+            ProgramInput::Resolved(ResolvedInput {
+                index: 3,
+                data: 2000u64.to_le_bytes().to_vec(),
+                input_type: ProgramInputType::Private,
+            }),
+            ProgramInput::Resolved(ResolvedInput {
+                index: 4,
+                data: (-2000i64).to_le_bytes().to_vec(),
+                input_type: ProgramInputType::Private,
+            }),
+            ProgramInput::Resolved(ResolvedInput {
+                index: 5,
+                data: "{\"attestation\":\"test\"}".as_bytes().to_vec(),
+                input_type: ProgramInputType::Private,
+            }),
+        ];
+        assert_eq!(inputs_parsed, expected_inputs);
+    }
+
+    #[test]
+    fn test_is_valid_number() {
+        let num = is_valid_number("1234567890abcdef");
+        assert!(num.is_none());
+        let num = is_valid_number("1234567890abcdefg");
+        assert!(num.is_none());
+        let num = is_valid_number("2.1");
+        assert!(num.is_some());
+        assert_eq!(num.unwrap(), NumberType::Float(2.1));
+        let num = is_valid_number("2000");
+        assert!(num.is_some());
+        assert_eq!(num.unwrap(), NumberType::Unsigned(2000));
+        let num = is_valid_number("-2000");
+        assert!(num.is_some());
+        assert_eq!(num.unwrap(), NumberType::Integer(-2000));
+    }
+
+    #[test]
+    fn test_execute_transform_cli_inputs() {
+        let input = CliInput {
+            input_type: "PublicData".to_string(),
+            data: "1234567890abcdef".to_string(),
+        };
+        let hex_input = CliInput {
+            input_type: "PublicData".to_string(),
+            data: "0x313233343536373839313061626364656667".to_string(),
+        };
+        let hex_input2 = CliInput {
+            input_type: "PublicData".to_string(),
+            data: "2.1".to_string(),
+        };
+        let hex_input3 = CliInput {
+            input_type: "PublicData".to_string(),
+            data: "2000".to_string(),
+        };
+        let hex_input4 = CliInput {
+            input_type: "PublicData".to_string(),
+            data: "-2000".to_string(),
+        };
+        let inputs = vec![input, hex_input, hex_input2, hex_input3, hex_input4];
+        let parsed_inputs = execute_transform_cli_inputs(inputs).unwrap();
+        assert_eq!(
+            parsed_inputs,
+            vec![
+                InputT::public("1234567890abcdef".as_bytes().to_vec()),
+                InputT::public("12345678910abcdefg".as_bytes().to_vec()),
+                InputT::public(2.1f64.to_le_bytes().to_vec()),
+                InputT::public(2000u64.to_le_bytes().to_vec()),
+                InputT::public((-2000i64).to_le_bytes().to_vec()),
+            ]
+        );
+    }
 }